+++ server/start.go
@@ -3,74 +3,53 @@ package server
 import (
 	"context"
 	"fmt"
-	"io"
-	"net"
-	"net/http"
-	"os"
-	"path/filepath"
-	"runtime/pprof"
 	"time"
 
-	errorsmod "cosmossdk.io/errors"
 	pruningtypes "cosmossdk.io/store/pruning/types"
-	abciserver "github.com/cometbft/cometbft/abci/server"
 	tcmd "github.com/cometbft/cometbft/cmd/cometbft/commands"
-	cmtcfg "github.com/cometbft/cometbft/config"
-	"github.com/cometbft/cometbft/node"
-	"github.com/cometbft/cometbft/p2p"
-	pvm "github.com/cometbft/cometbft/privval"
-	"github.com/cometbft/cometbft/proxy"
-	rpcclient "github.com/cometbft/cometbft/rpc/client"
-	"github.com/cometbft/cometbft/rpc/client/local"
-	cmttypes "github.com/cometbft/cometbft/types"
 	dbm "github.com/cosmos/cosmos-db"
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/flags"
-	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/crypto/keyring"
 	"github.com/cosmos/cosmos-sdk/server"
-	"github.com/cosmos/cosmos-sdk/server/api"
 	serverconfig "github.com/cosmos/cosmos-sdk/server/config"
-	servergrpc "github.com/cosmos/cosmos-sdk/server/grpc"
-	servercmtlog "github.com/cosmos/cosmos-sdk/server/log"
 	"github.com/cosmos/cosmos-sdk/server/types"
-	"github.com/cosmos/cosmos-sdk/telemetry"
-	genutiltypes "github.com/cosmos/cosmos-sdk/x/genutil/types"
-	"github.com/cosmos/evm/indexer"
 	cosmosevmserverconfig "github.com/cosmos/evm/server/config"
 	srvflags "github.com/cosmos/evm/server/flags"
-	cosmosevmtypes "github.com/cosmos/evm/types"
-	ethmetricsexp "github.com/ethereum/go-ethereum/metrics/exp"
 	"github.com/spf13/cobra"
 	"golang.org/x/sync/errgroup"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/credentials/insecure"
-
-	ethdebug "github.com/zeta-chain/node/rpc/namespaces/ethereum/debug"
 )
 
-// DBOpener is a function to open `application.db`, potentially with customized options.
-type DBOpener func(opts types.AppOptions, rootDir string, backend dbm.BackendType) (dbm.DB, error)
+const (
+	KeyIsDevnet                 = "is-devnet"
+	KeyNewChainID               = "new-chain-ID"
+	KeyValidatorConsensusAddr   = "validator-consensus-address"
+	KeyValidatorConsensusPubkey = "validator-consensus-pubkey"
+	KeyOperatorAddress          = "operator-address"
+	FlagSkipConfirmation        = "skip-confirmation"
+)
 
-// StartOptions defines options that can be customized in `StartCmd`
-type StartOptions struct {
-	AppCreator      types.AppCreator
-	DefaultNodeHome string
-	DBOpener        DBOpener
-}
-
-// NewDefaultStartOptions use the default db opener provided in tm-db.
-func NewDefaultStartOptions(appCreator types.AppCreator, defaultNodeHome string) StartOptions {
-	return StartOptions{
-		AppCreator:      appCreator,
-		DefaultNodeHome: defaultNodeHome,
-		DBOpener:        cosmosevmserverconfig.OpenDB,
-	}
+type StartCmdOptions struct {
+	DBOpener            func(rootDir string, backendType dbm.BackendType) (dbm.DB, error)
+	PostSetup           func(svrCtx *server.Context, clientCtx client.Context, ctx context.Context, g *errgroup.Group) error
+	PostSetupStandalone func(svrCtx *server.Context, clientCtx client.Context, ctx context.Context, g *errgroup.Group) error
+	AddFlags            func(cmd *cobra.Command)
+	StartCommandHandler func(svrCtx *server.Context, clientCtx client.Context, appCreator types.AppCreator, withCmt bool, opts StartCmdOptions) error
 }
 
 // StartCmd runs the service passed in, either stand-alone or in-process with
 // CometBFT.
-func StartCmd(opts StartOptions) *cobra.Command {
+func StartCmd(appCreator types.AppCreator, defaultNodeHome string) *cobra.Command {
+	return StartCmdWithOptions(appCreator, defaultNodeHome, StartCmdOptions{
+		DBOpener:            openDB,
+		StartCommandHandler: start,
+	})
+}
+
+func StartCmdWithOptions(appCreator types.AppCreator, defaultNodeHome string, opts StartCmdOptions) *cobra.Command {
+	if opts.DBOpener == nil || opts.StartCommandHandler == nil {
+		panic("DBOpener and StartCommandHandler must be provided")
+	}
 	cmd := &cobra.Command{
 		Use:   "start",
 		Short: "Run the full node",
@@ -116,15 +95,9 @@ which accepts a path for the resulting pprof file.
 				return err
 			}
 
-			withTM, _ := cmd.Flags().GetBool(srvflags.WithCometBFT)
-			if !withTM {
-				serverCtx.Logger.Info("starting ABCI without Tendermint")
-				return startStandAlone(serverCtx, opts)
-			}
-
 			serverCtx.Logger.Info("Unlocking keyring")
 
-			// fire unlock precess for keyring
+			// fire unlock process for keyring
 			keyringBackend, _ := cmd.Flags().GetString(flags.FlagKeyringBackend)
 			if keyringBackend == keyring.BackendFile {
 				_, err = clientCtx.Keyring.List()
@@ -137,18 +110,6 @@ which accepts a path for the resulting pprof file.
 			if val, err := cmd.Flags().GetBool(FlagSkipConfigOverwrite); err == nil {
 				skipOverwrite = val
 			}
-			//zevmChainID, err := genesisChainID(serverCtx.Config.GenesisFile())
-			//if err != nil {
-			//	return errorsmod.Wrapf(err, "failed to get genesis chain ID from genesis file")
-			//}
-			//
-			//// Cannot skip over writing the config file for ZetaChain mainnet
-			//if zevmChainID == chains.ZetaChainMainnet.ChainId && skipOverwrite {
-			//	return fmt.Errorf(
-			//		"config overwrite is required for ZetaChain mainnet , please run the command without the --%s flag",
-			//		FlagSkipConfigOverwrite,
-			//	)
-			//}
 
 			if !skipOverwrite {
 				err := overWriteConfig(cmd)
@@ -156,34 +117,26 @@ which accepts a path for the resulting pprof file.
 					return fmt.Errorf("failed to overwrite config: %w", err)
 				}
 			}
-			serverCtx.Logger.Info("starting ABCI with CometBFT")
 
-			// amino is needed here for backwards compatibility of REST routes
-			if err := startInProcess(serverCtx, clientCtx, opts); err != nil {
-				return err
-			}
+			withCmt, _ := cmd.Flags().GetBool(srvflags.WithCometBFT)
+
+			err = opts.StartCommandHandler(serverCtx, clientCtx, appCreator, withCmt, opts)
 
 			serverCtx.Logger.Debug("received quit signal")
-			// TODO: why is this check here? Should not make sense since err is checked above
-			if err != nil {
-				serverCtx.Logger.Error(fmt.Sprintf("error on quit: %s", err.Error()))
-			}
-
-			return nil
+			return err
 		},
 	}
 
 	cmd.Flags().
-		Bool(FlagSkipConfigOverwrite, false, "Skip running the config configuration overwrite handler.This is used for testing purposes only and skips using the default timeouts hardcoded and uses the config file instead")
-	cmd.Flags().String(flags.FlagHome, opts.DefaultNodeHome, "The application home directory")
+		Bool(FlagSkipConfigOverwrite, false, "Skip running the config configuration overwrite handler. This is used for testing purposes only and skips using the default timeouts hardcoded and uses the config file instead")
+	cmd.Flags().String(flags.FlagHome, defaultNodeHome, "The application home directory")
 	cmd.Flags().Bool(srvflags.WithCometBFT, true, "Run abci app embedded in-process with CometBFT")
 	cmd.Flags().String(srvflags.Address, "tcp://0.0.0.0:26658", "Listen address")
 	cmd.Flags().String(srvflags.Transport, "socket", "Transport protocol: socket, grpc")
 	cmd.Flags().String(srvflags.TraceStore, "", "Enable KVStore tracing to an output file")
 	cmd.Flags().
 		String(server.FlagMinGasPrices, "", "Minimum gas prices to accept for transactions; Any fee in a tx must meet this minimum (e.g. 20000000000azeta)")
-
-	//nolint:lll
+	cmd.Flags().Duration(server.FlagShutdownGrace, 3*time.Second, "On Shutdown, duration to wait for resource clean up")
 	cmd.Flags().
 		IntSlice(server.FlagUnsafeSkipUpgrades, []int{}, "Skip a set of upgrade heights to continue the old binary")
 	cmd.Flags().
@@ -200,7 +153,6 @@ which accepts a path for the resulting pprof file.
 	cmd.Flags().
 		Uint64(server.FlagPruningInterval, 0, "Height interval at which pruned heights are removed from disk (ignored if pruning is not 'custom')")
 
-	//nolint:lll
 	cmd.Flags().Uint(server.FlagInvCheckPeriod, 0, "Assert registered invariants every N blocks")
 	cmd.Flags().
 		Uint64(server.FlagMinRetainBlocks, 0, "Minimum block height offset during ABCI commit to prune CometBFT blocks")
@@ -231,15 +183,10 @@ which accepts a path for the resulting pprof file.
 	cmd.Flags().
 		Uint64(srvflags.JSONRPCGasCap, cosmosevmserverconfig.DefaultGasCap, "Sets a cap on gas that can be used in eth_call/estimateGas unit is aatom (0=infinite)")
 
-	//nolint:lll
 	cmd.Flags().
 		Bool(srvflags.JSONRPCAllowInsecureUnlock, cosmosevmserverconfig.DefaultJSONRPCAllowInsecureUnlock, "Allow insecure account unlocking when account-related RPCs are exposed by http")
-
-	//nolint:lll
 	cmd.Flags().
 		Float64(srvflags.JSONRPCTxFeeCap, cosmosevmserverconfig.DefaultTxFeeCap, "Sets a cap on transaction fee that can be sent via the RPC APIs (1 = default 1 evmos)")
-
-	//nolint:lll
 	cmd.Flags().
 		Int32(srvflags.JSONRPCFilterCap, cosmosevmserverconfig.DefaultFilterCap, "Sets the global cap for total number of filters that can be created")
 	cmd.Flags().
@@ -251,7 +198,6 @@ which accepts a path for the resulting pprof file.
 	cmd.Flags().
 		Bool(srvflags.JSONRPCAllowUnprotectedTxs, cosmosevmserverconfig.DefaultAllowUnprotectedTxs, "Allow for unprotected (non EIP155 signed) transactions to be submitted via the node's RPC when the global parameter is disabled")
 
-	//nolint:lll
 	cmd.Flags().
 		Int32(srvflags.JSONRPCLogsCap, cosmosevmserverconfig.DefaultLogsCap, "Sets the max number of results can be returned from single `eth_getLogs` query")
 	cmd.Flags().
@@ -259,22 +205,15 @@ which accepts a path for the resulting pprof file.
 	cmd.Flags().
 		Int(srvflags.JSONRPCMaxOpenConnections, cosmosevmserverconfig.DefaultMaxOpenConnections, "Sets the maximum number of simultaneous connections for the server listener")
 
-	//nolint:lll
 	cmd.Flags().Bool(srvflags.JSONRPCEnableIndexer, false, "Enable the custom tx indexer for json-rpc")
 	cmd.Flags().Bool(srvflags.JSONRPCEnableMetrics, false, "Define if EVM rpc metrics server should be enabled")
 
 	cmd.Flags().
 		String(srvflags.EVMTracer, cosmosevmserverconfig.DefaultEVMTracer, "the EVM tracer type to collect execution traces from the EVM transaction execution (json|struct|access_list|markdown)")
-
-	//nolint:lll
 	cmd.Flags().
 		Uint64(srvflags.EVMMaxTxGasWanted, cosmosevmserverconfig.DefaultMaxTxGasWanted, "the gas wanted for each eth tx returned in ante handler in check tx mode")
-
-	//nolint:lll
 	cmd.Flags().
 		Bool(srvflags.EVMEnablePreimageRecording, cosmosevmserverconfig.DefaultEnablePreimageRecording, "Enables tracking of SHA3 preimages in the EVM (not implemented yet)")
-
-	//nolint:lll
 	cmd.Flags().
 		Uint64(srvflags.EVMChainID, cosmosevmserverconfig.DefaultEVMChainID, "the EIP-155 compatible replay protection chain ID")
 
@@ -284,475 +223,14 @@ which accepts a path for the resulting pprof file.
 	cmd.Flags().Uint64(server.FlagStateSyncSnapshotInterval, 0, "State sync snapshot interval")
 	cmd.Flags().Uint32(server.FlagStateSyncSnapshotKeepRecent, 2, "State sync snapshot to keep")
 
+	cmd.Flags().Bool(KeyIsDevnet, false, "Enable devnet mode to fork from existing state")
+	cmd.Flags().String(KeyNewChainID, "", "New chain ID to use when running in devnet mode")
+
+	if opts.AddFlags != nil {
+		opts.AddFlags(cmd)
+	}
+
 	// add support for all CometBFT-specific command line options
 	tcmd.AddNodeFlags(cmd)
 	return cmd
-}
-
-// startStandAlone starts an ABCI server in stand-alone mode.
-// Parameters:
-// - svrCtx: The context object that holds server configurations, logger, and other stateful information.
-// - opts: Options for starting the server, including functions for creating the application and opening the database.
-func startStandAlone(svrCtx *server.Context, opts StartOptions) error {
-	addr := svrCtx.Viper.GetString(srvflags.Address)
-	transport := svrCtx.Viper.GetString(srvflags.Transport)
-	home := svrCtx.Viper.GetString(flags.FlagHome)
-
-	db, err := opts.DBOpener(svrCtx.Viper, home, server.GetAppDBBackend(svrCtx.Viper))
-	if err != nil {
-		return err
-	}
-
-	defer func() {
-		if err := db.Close(); err != nil {
-			svrCtx.Logger.Error("error closing db", "error", err.Error())
-		}
-	}()
-
-	traceWriterFile := svrCtx.Viper.GetString(srvflags.TraceStore)
-	traceWriter, err := openTraceWriter(traceWriterFile)
-	if err != nil {
-		return err
-	}
-
-	app := opts.AppCreator(svrCtx.Logger, db, traceWriter, svrCtx.Viper)
-
-	config, err := cosmosevmserverconfig.GetConfig(svrCtx.Viper)
-	if err != nil {
-		svrCtx.Logger.Error("failed to get server config", "error", err.Error())
-		return err
-	}
-
-	if err := config.ValidateBasic(); err != nil {
-		svrCtx.Logger.Error("invalid server config", "error", err.Error())
-		return err
-	}
-
-	_, err = startTelemetry(config)
-	if err != nil {
-		return err
-	}
-
-	cmtApp := server.NewCometABCIWrapper(app)
-	svr, err := abciserver.NewServer(addr, transport, cmtApp)
-	if err != nil {
-		return fmt.Errorf("error creating listener: %v", err)
-	}
-
-	svr.SetLogger(servercmtlog.CometLoggerWrapper{Logger: svrCtx.Logger.With("server", "abci")})
-	g, ctx := getCtx(svrCtx, false)
-
-	g.Go(func() error {
-		if err := svr.Start(); err != nil {
-			svrCtx.Logger.Error("failed to start out-of-process ABCI server", "err", err)
-			return err
-		}
-
-		// Wait for the calling process to be canceled or close the provided context,
-		// so we can gracefully stop the ABCI server.
-		<-ctx.Done()
-		svrCtx.Logger.Info("stopping the ABCI server...")
-		return svr.Stop()
-	})
-
-	return g.Wait()
-}
-
-// legacyAminoCdc is used for the legacy REST API
-func startInProcess(svrCtx *server.Context, clientCtx client.Context, opts StartOptions) (err error) {
-	cfg := svrCtx.Config
-	home := cfg.RootDir
-	logger := svrCtx.Logger
-	g, ctx := getCtx(svrCtx, true)
-
-	if cpuProfile := svrCtx.Viper.GetString(srvflags.CPUProfile); cpuProfile != "" {
-		fp, err := ethdebug.ExpandHome(cpuProfile)
-		if err != nil {
-			svrCtx.Logger.Debug("failed to get filepath for the CPU profile file", "error", err.Error())
-			return err
-		}
-
-		f, err := os.Create(fp) //#nosec G304 forked code
-		if err != nil {
-			return err
-		}
-
-		svrCtx.Logger.Info("starting CPU profiler", "profile", cpuProfile)
-		if err := pprof.StartCPUProfile(f); err != nil {
-			return err
-		}
-
-		defer func() {
-			svrCtx.Logger.Info("stopping CPU profiler", "profile", cpuProfile)
-			pprof.StopCPUProfile()
-			if err := f.Close(); err != nil {
-				logger.Error("failed to close CPU profiler file", "error", err.Error())
-			}
-		}()
-	}
-
-	db, err := opts.DBOpener(svrCtx.Viper, home, server.GetAppDBBackend(svrCtx.Viper))
-	if err != nil {
-		logger.Error("failed to open DB", "error", err.Error())
-		return err
-	}
-
-	defer func() {
-		if err := db.Close(); err != nil {
-			svrCtx.Logger.With("error", err).Error("error closing db")
-		}
-	}()
-
-	traceWriterFile := svrCtx.Viper.GetString(srvflags.TraceStore)
-	traceWriter, err := openTraceWriter(traceWriterFile)
-	if err != nil {
-		logger.Error("failed to open trace writer", "error", err.Error())
-		return err
-	}
-
-	config, err := cosmosevmserverconfig.GetConfig(svrCtx.Viper)
-	if err != nil {
-		logger.Error("failed to get server config", "error", err.Error())
-		return err
-	}
-
-	if err := config.ValidateBasic(); err != nil {
-		logger.Error("invalid server config", "error", err.Error())
-		return err
-	}
-
-	app := opts.AppCreator(svrCtx.Logger, db, traceWriter, svrCtx.Viper)
-
-	nodeKey, err := p2p.LoadOrGenNodeKey(cfg.NodeKeyFile())
-	if err != nil {
-		logger.Error("failed load or gen node key", "error", err.Error())
-		return err
-	}
-
-	genDocProvider := GenDocProvider(cfg)
-
-	var (
-		tmNode   *node.Node
-		gRPCOnly = svrCtx.Viper.GetBool(srvflags.GRPCOnly)
-	)
-
-	if gRPCOnly {
-		logger.Info("starting node in query only mode; CometBFT is disabled")
-		config.GRPC.Enable = true
-		config.JSONRPC.EnableIndexer = false
-	} else {
-		logger.Info("starting node with ABCI CometBFT in-process")
-
-		cmtApp := server.NewCometABCIWrapper(app)
-		tmNode, err = node.NewNode(
-			cfg,
-			pvm.LoadOrGenFilePV(cfg.PrivValidatorKeyFile(), cfg.PrivValidatorStateFile()),
-			nodeKey,
-			proxy.NewLocalClientCreator(cmtApp),
-			genDocProvider,
-			cmtcfg.DefaultDBProvider,
-			node.DefaultMetricsProvider(cfg.Instrumentation),
-			servercmtlog.CometLoggerWrapper{Logger: svrCtx.Logger.With("server", "node")},
-		)
-		if err != nil {
-			logger.Error("failed init node", "error", err.Error())
-			return err
-		}
-
-		if err := tmNode.Start(); err != nil {
-			logger.Error("failed start CometBFT server", "error", err.Error())
-			return err
-		}
-
-		defer func() {
-			if tmNode.IsRunning() {
-				_ = tmNode.Stop()
-			}
-		}()
-	}
-
-	// Add the tx service to the gRPC router. We only need to register this
-	// service if API or gRPC or JSONRPC is enabled, and avoid doing so in the general
-	// case, because it spawns a new local CometBFT RPC client.
-	if (config.API.Enable || config.GRPC.Enable || config.JSONRPC.Enable || config.JSONRPC.EnableIndexer) &&
-		tmNode != nil {
-		clientCtx = clientCtx.WithClient(local.New(tmNode))
-
-		app.RegisterTxService(clientCtx)
-		app.RegisterTendermintService(clientCtx)
-		app.RegisterNodeService(clientCtx, config.Config)
-	}
-
-	metrics, err := startTelemetry(config)
-	if err != nil {
-		return err
-	}
-
-	// Enable metrics if JSONRPC is enabled and --metrics is passed
-	// Flag not added in config to avoid user enabling in config without passing in CLI
-	if config.JSONRPC.Enable && svrCtx.Viper.GetBool(srvflags.JSONRPCEnableMetrics) {
-		ethmetricsexp.Setup(config.JSONRPC.MetricsAddress)
-	}
-
-	var idxer cosmosevmtypes.EVMTxIndexer
-	if config.JSONRPC.EnableIndexer {
-		idxDB, err := OpenIndexerDB(home, server.GetAppDBBackend(svrCtx.Viper))
-		if err != nil {
-			logger.Error("failed to open evm indexer DB", "error", err.Error())
-			return err
-		}
-
-		idxLogger := svrCtx.Logger.With("indexer", "evm")
-		idxer = indexer.NewKVIndexer(idxDB, idxLogger, clientCtx)
-		indexerService := NewEVMIndexerService(idxer, clientCtx.Client.(rpcclient.Client))
-		indexerService.SetLogger(servercmtlog.CometLoggerWrapper{Logger: idxLogger})
-
-		g.Go(func() error {
-			return indexerService.Start()
-		})
-	}
-
-	if config.API.Enable || config.JSONRPC.Enable {
-		genDoc, err := genDocProvider()
-		if err != nil {
-			return err
-		}
-
-		clientCtx = clientCtx.
-			WithHomeDir(home).
-			WithChainID(genDoc.ChainID)
-	}
-
-	grpcSrv, clientCtx, err := startGrpcServer(ctx, svrCtx, clientCtx, g, config.GRPC, app)
-	if err != nil {
-		return err
-	}
-	if grpcSrv != nil {
-		defer grpcSrv.GracefulStop()
-	}
-
-	apiSrv := startAPIServer(ctx, svrCtx, clientCtx, g, config.Config, app, grpcSrv, metrics)
-
-	if apiSrv != nil {
-		defer apiSrv.Close()
-	}
-
-	clientCtx, httpSrv, httpSrvDone, err := startJSONRPCServer(
-		svrCtx,
-		clientCtx,
-		g,
-		config,
-		genDocProvider,
-		cfg.RPC.ListenAddress,
-		idxer,
-	)
-	if httpSrv != nil {
-		defer func() {
-			shutdownCtx, cancelFn := context.WithTimeout(context.Background(), 10*time.Second)
-			defer cancelFn()
-			if err := httpSrv.Shutdown(shutdownCtx); err != nil {
-				logger.Error("HTTP server shutdown produced a warning", "error", err.Error())
-			} else {
-				logger.Info("HTTP server shut down, waiting 5 sec")
-				select {
-				case <-time.Tick(5 * time.Second):
-				case <-httpSrvDone:
-				}
-			}
-		}()
-	}
-
-	// At this point it is safe to block the process if we're in query only mode as
-	// we do not need to start Rosetta or handle any CometBFT related processes.
-	if gRPCOnly {
-		// wait for signal capture and gracefully return
-		// we are guaranteed to be waiting for the "ListenForQuitSignals" goroutine.
-		return g.Wait()
-	}
-
-	// wait for signal capture and gracefully return
-	// we are guaranteed to be waiting for the "ListenForQuitSignals" goroutine.
-	return g.Wait()
-}
-
-// OpenIndexerDB opens the custom eth indexer db, using the same db backend as the main app
-func OpenIndexerDB(rootDir string, backendType dbm.BackendType) (dbm.DB, error) {
-	dataDir := filepath.Join(rootDir, "data")
-	return dbm.NewDB("evmindexer", backendType, dataDir)
-}
-
-// openTraceWriter opens a trace writer if a trace store file is specified.
-// Parameters:
-// - traceWriterFile: The path to the trace store file. If this is an empty string, no file will be opened.
-func openTraceWriter(traceWriterFile string) (w io.Writer, err error) {
-	if traceWriterFile == "" {
-		return
-	}
-
-	filePath := filepath.Clean(traceWriterFile)
-	return os.OpenFile(
-		filePath,
-		os.O_WRONLY|os.O_APPEND|os.O_CREATE,
-		0o600,
-	)
-}
-
-func startTelemetry(cfg cosmosevmserverconfig.Config) (*telemetry.Metrics, error) {
-	if !cfg.Telemetry.Enabled {
-		return nil, nil
-	}
-	return telemetry.New(cfg.Telemetry)
-}
-
-func getCtx(svrCtx *server.Context, block bool) (*errgroup.Group, context.Context) {
-	ctx, cancelFn := context.WithCancel(context.Background())
-	g, ctx := errgroup.WithContext(ctx)
-	// listen for quit signals so the calling parent process can gracefully exit
-	server.ListenForQuitSignals(g, block, cancelFn, svrCtx.Logger)
-	return g, ctx
-}
-
-// startGrpcServer starts a gRPC server based on the provided configuration.
-func startGrpcServer(
-	ctx context.Context,
-	svrCtx *server.Context,
-	clientCtx client.Context,
-	g *errgroup.Group,
-	config serverconfig.GRPCConfig,
-	app types.Application,
-) (*grpc.Server, client.Context, error) {
-	if !config.Enable {
-		// return grpcServer as nil if gRPC is disabled
-		return nil, clientCtx, nil
-	}
-	_, _, err := net.SplitHostPort(config.Address)
-	if err != nil {
-		return nil, clientCtx, errorsmod.Wrapf(err, "invalid grpc address %s", config.Address)
-	}
-
-	maxSendMsgSize := config.MaxSendMsgSize
-	if maxSendMsgSize == 0 {
-		maxSendMsgSize = serverconfig.DefaultGRPCMaxSendMsgSize
-	}
-
-	maxRecvMsgSize := config.MaxRecvMsgSize
-	if maxRecvMsgSize == 0 {
-		maxRecvMsgSize = serverconfig.DefaultGRPCMaxRecvMsgSize
-	}
-
-	// if gRPC is enabled, configure gRPC client for gRPC gateway and json-rpc
-	grpcClient, err := grpc.NewClient(
-		config.Address,
-		grpc.WithTransportCredentials(insecure.NewCredentials()),
-		grpc.WithDefaultCallOptions(
-			grpc.ForceCodec(codec.NewProtoCodec(clientCtx.InterfaceRegistry).GRPCCodec()),
-			grpc.MaxCallRecvMsgSize(maxRecvMsgSize),
-			grpc.MaxCallSendMsgSize(maxSendMsgSize),
-		),
-	)
-	if err != nil {
-		return nil, clientCtx, err
-	}
-	// Set `GRPCClient` to `clientCtx` to enjoy concurrent grpc query.
-	// only use it if gRPC server is enabled.
-	clientCtx = clientCtx.WithGRPCClient(grpcClient)
-	svrCtx.Logger.Debug("gRPC client assigned to client context", "address", config.Address)
-
-	grpcSrv, err := servergrpc.NewGRPCServer(clientCtx, app, config)
-	if err != nil {
-		return nil, clientCtx, err
-	}
-
-	// Start the gRPC server in a goroutine. Note, the provided ctx will ensure
-	// that the server is gracefully shut down.
-	g.Go(func() error {
-		return servergrpc.StartGRPCServer(ctx, svrCtx.Logger.With("module", "grpc-server"), config, grpcSrv)
-	})
-	return grpcSrv, clientCtx, nil
-}
-
-// startAPIServer starts an API server based on the provided configuration and application context.
-// Parameters:
-// - ctx: The context used for managing the server's lifecycle, allowing for graceful shutdown.
-// - svrCtx: The server context containing configuration, logger, and other stateful components.
-// - clientCtx: The client context, which provides necessary information for API operations.
-// - g: An errgroup.Group for managing goroutines and handling errors concurrently.
-// - svrCfg: The server configuration that specifies whether the API server is enabled and other settings.
-// - app: The application instance that registers API routes.
-// - grpcSrv: A pointer to the gRPC server, which may be used by the API server.
-// - metrics: A telemetry metrics instance for monitoring API server performance.
-func startAPIServer(
-	ctx context.Context,
-	svrCtx *server.Context,
-	clientCtx client.Context,
-	g *errgroup.Group,
-	svrCfg serverconfig.Config,
-	app types.Application,
-	grpcSrv *grpc.Server,
-	metrics *telemetry.Metrics,
-) *api.Server {
-	if !svrCfg.API.Enable {
-		return nil
-	}
-
-	apiSrv := api.New(clientCtx, svrCtx.Logger.With("server", "api"), grpcSrv)
-	app.RegisterAPIRoutes(apiSrv, svrCfg.API)
-
-	if svrCfg.Telemetry.Enabled {
-		apiSrv.SetTelemetry(metrics)
-	}
-
-	g.Go(func() error {
-		return apiSrv.Start(ctx, svrCfg)
-	})
-	return apiSrv
-}
-
-// startJSONRPCServer starts a JSON-RPC server based on the provided configuration.
-// Parameters:
-// - svrCtx: The server context containing configuration, logger, and stateful components.
-// - clientCtx: The client context, which may be updated with additional chain information.
-// - g: An errgroup.Group to manage concurrent goroutines and error handling.
-// - config: The server configuration that specifies whether the JSON-RPC server is enabled and other settings.
-// - genDocProvider: A function that provides the Genesis document, used to retrieve the chain ID.
-// - cmtRPCAddr: The address of the CometBFT RPC server for WebSocket connections.
-// - idxer: The EVM transaction indexer for indexing transactions.
-func startJSONRPCServer(
-	svrCtx *server.Context,
-	clientCtx client.Context,
-	g *errgroup.Group,
-	config cosmosevmserverconfig.Config,
-	genDocProvider node.GenesisDocProvider,
-	cmtRPCAddr string,
-	idxer cosmosevmtypes.EVMTxIndexer,
-) (ctx client.Context, httpSrv *http.Server, httpSrvDone chan struct{}, err error) {
-	ctx = clientCtx
-	if !config.JSONRPC.Enable {
-		return
-	}
-
-	genDoc, err := genDocProvider()
-	if err != nil {
-		return ctx, httpSrv, httpSrvDone, err
-	}
-
-	ctx = clientCtx.WithChainID(genDoc.ChainID)
-	cmtEndpoint := "/websocket"
-	g.Go(func() error {
-		httpSrv, httpSrvDone, err = StartJSONRPC(svrCtx, clientCtx, cmtRPCAddr, cmtEndpoint, &config, idxer)
-		return err
-	})
-	return
-}
-
-// GenDocProvider returns a function which returns the genesis doc from the genesis file.
-func GenDocProvider(cfg *cmtcfg.Config) func() (*cmttypes.GenesisDoc, error) {
-	return func() (*cmttypes.GenesisDoc, error) {
-		appGenesis, err := genutiltypes.AppGenesisFromFile(cfg.GenesisFile())
-		if err != nil {
-			return nil, err
-		}
-
-		return appGenesis.ToGenesisDoc()
-	}
 }