+++ changelog.md
@@ -68,6 +68,7 @@ Also EnableSolanaAddressLookupTable feature flag should be set.
 * [4362](https://github.com/zeta-chain/node/pull/4362) - shutdown zetaclient if zetacore is syncing
 * [4378](https://github.com/zeta-chain/node/pull/4378) - add number of connected peers to tss keygen log
 * [4387](https://github.com/zeta-chain/node/pull/4387) - add new dry zetaclient to localnet
+* [4419](https://github.com/zeta-chain/node/pull/4419) - add testnet command to create a fork from existing node data
 
 ### Tests
 
@@ -76,6 +77,7 @@ Also EnableSolanaAddressLookupTable feature flag should be set.
 * [4357](https://github.com/zeta-chain/node/pull/4357) - json rpc checks during e2e and upgrade tests
 * [4395](https://github.com/zeta-chain/node/pull/4395) - update e2e to set higher timeouts when running stress test
 * [4408](https://github.com/zeta-chain/node/pull/4408) - add chaos mode capabilities to ZetaClient localnet
+
 
 ### Refactor